Added Pillar-Button parameters and skin claim.
Added the parameters for the pillar button but the default values for each are still required (unsure what the values for them are so left them blank).

Also deleted the claim that more skins are coming later (saves against possible criticism if it takes awhile to get done).

Also specified specific versions of python in the installation guide as recent versions have issues (found this when switching laptops and installing animal ai was an issue).